refactor(ui): one Settings window with tabs, not two windows
The two-window split was justified by "a TabControl does not auto-size, and
SettingsForm auto-sizes around a Discord group that swaps a tall panel for a
short one". That was true in WinForms. The WPF window has a fixed height and a
ScrollViewer, so the constraint died with the port — the justification was
carried over without being re-examined.

It was costing something real: two Save buttons with different semantics (one
gated on Discord and restarting the app, one immediate), a window that opened
modally from Settings and modelessly from the tray, and a second top-level
window to lose behind another app.

Now one window, two tabs — Connection and Overlay — and one Save that writes
the OBS section, auto-start and the Streamkit section together, then re-pushes
the overlay URL. The tray's "Overlay appearance..." entry still goes straight
there: it opens the same window on the Overlay tab.

OverlayAppearanceWindow is gone, along with the strings that only existed to
title it and report its separate save.
